requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
brotli==1.0.9
cached-property==1.5.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
cycler==0.11.0
cytoolz==0.12.0
django==4.1
enum34==1.1.10
filelock==3.12.2
fonttools==4.25.0
fsspec==2023.6.0
ftfy==6.1.1
gunicorn==21.2.0
h5py==3.7.0
huggingface-hub==0.16.4
humanize==4.8.0
idna==3.4
imagecodecs-lite==2019.12.3
imageio==2.31.1
importlib-resources==6.0.1
intel-openmp==2023.2.0
jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
locket==1.0.0
markdown-it-py==3.0.0
markupsafe==2.1.3
matplotlib==3.5.3
mdurl==0.1.2
mpmath==1.3.0
munkres==1.1.4
natsort==8.4.0
networkx==3.1
nibabel==5.1.0
numpy==1.24.4
open-clip-torch==2.20.0
opencv-python==4.7.0.68
packaging==23.1
pandas==2.0.3
partd==1.4.0
pathlib==1.0.1
pillow==10.0.0
pip==23.2.1
platformdirs==3.10.0
ply==3.11
pooch==1.7.0
protobuf==3.20.3
pygments==2.16.1
pyparsing==3.1.1
pyqt5-sip==12.11.0
pysocks==1.7.1
python-dateutil==2.8.2
pywavelets==1.4.1
pyyaml==6.0.1
regex==2023.8.8
requests==2.31.0
rich==13.5.2
safetensors==0.3.2
scikit-image==0.18.1
scipy==1.8.1
sentencepiece==0.1.99
setuptools==68.0.0
shellingham==1.5.3
simpleitk==2.2.1
sip==6.6.2
six==1.16.0
sqlparse==0.4.3
sympy==1.12
tifffile==2019.7.26.2
timm==0.9.5
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchio==0.19.1
torchvision==0.15.2
tornado==6.2
tqdm==4.66.1
transformers==4.31.0
typer==0.9.0
typing-extensions==4.7.1
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
wheel==0.38.4
win_inet_pton==1.1.0
wrapt==1.15.0
yacs==0.1.8
zipp==3.16.2
google-auth
google-auth-httplib2